# backend/src/swagger/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "DukaPay API",
    "version": "1.0.0",
    "description": "Backend API for DukaPay lending, scoring, remittance, and indexer flows."
  },
  "servers": [
    {
      "url": "/api",
      "description": "Legacy API base path"
    },
    {
      "url": "/api/v1",
      "description": "Versioned API base path"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "UnsignedTransactionResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "unsignedTxXdr": {
            "type": "string"
          },
          "networkPassphrase": {
            "type": "string"
          }
        },
        "required": [
          "success",
          "unsignedTxXdr",
          "networkPassphrase"
        ]
      },
...
      "SubmittedTransactionResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "txHash": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "resultXdr": {
            "type": "string"
          }
        },
        "required": [
          "success",
          "txHash",
          "status"
        ]
      },
      "PoolStats": {
        "type": "object",
        "properties": {
          "totalDeposits": {
            "type": "number"
          },
          "totalOutstanding": {
            "type": "number"
          },
          "utilizationRate": {
            "type": "number"
          },
          "apy": {
            "type": "number"
          },
          "activeLoansCount": {
            "type": "integer"
          }
        },
        "required": [
          "totalDeposits",
          "totalOutstanding",
          "utilizationRate",
          "apy",
          "activeLoansCount"
        ]
      },
      "PoolStatsResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "$ref": "#/components/schemas/PoolStats"
          }
        },
        "required": [
          "success",
          "data"
        ]
      },
      "DepositorPortfolio": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string"
          },
          "depositAmount": {
            "type": "number"
          },
          "sharePercent": {
            "type": "number"
          },
          "estimatedYield": {
            "type": "number"
          },
          "apy": {
            "type": "number"
          },
          "firstDepositAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": [
          "address",
          "depositAmount",
          "sharePercent",
          "estimatedYield",
          "apy"
        ]
      },
      "DepositorPortfolioResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "$ref": "#/components/schemas/DepositorPortfolio"
          }
        },
        "required": [
          "success",
          "data"
        ]
      },
...
    }
  },
  "paths": {
...
    "/pool/analytics": {
      "get": {
        "summary": "Get aggregate pool analytics",
        "description": "Returns high-level lending-pool analytics: total deposits and withdrawals, yield distributed, loans issued, total volume and the number of active agents. Public endpoint; cached for 300 seconds.\n",
        "tags": [
          "Pool"
        ],
        "responses": {
          "200": {
            "description": "Pool analytics retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "analytics",
                    "source"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "analytics": {
                      "type": "object",
                      "required": [
                        "totalDeposits",
                        "totalWithdrawals",
                        "totalYieldDistributed",
                        "totalLoansIssued",
                        "totalVolume",
                        "activeAgents",
                        "updatedAt"
                      ],
                      "properties": {
                        "totalDeposits": {
                          "type": "number"
                        },
                        "totalWithdrawals": {
                          "type": "number"
                        },
                        "totalYieldDistributed": {
                          "type": "number"
                        },
                        "totalLoansIssued": {
                          "type": "integer"
                        },
                        "totalVolume": {
                          "type": "number"
                        },
                        "activeAgents": {
                          "type": "integer"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    },
                    "source": {
                      "type": "string",
                      "enum": [
                        "cache",
                        "database"
                      ]
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/pool/agent/dashboard/{agentAddress}": {
      "get": {
        "summary": "Get an agent's lending dashboard",
        "description": "Returns a single agent's float balance, collateral balance, active state, and number of active loans. Restricted to the dashboard owner (agent/lender), admins and auditors under tenant isolation; cached for the active scope for 30 seconds.\n",
        "tags": [
          "Pool"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "agentAddress",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Stellar public key of the agent"
          }
        ],
        "responses": {
          "200": {
            "description": "Agent dashboard retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "dashboard",
                    "source"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "dashboard": {
                      "type": "object",
                      "required": [
                        "agentAddress",
                        "floatBalance",
                        "collateralBalance",
                        "isActive",
                        "activeLoans",
                        "updatedAt"
                      ],
                      "properties": {
                        "agentAddress": {
                          "type": "string"
                        },
                        "floatBalance": {
                          "type": "number"
                        },
                        "collateralBalance": {
                          "type": "number"
                        },
                        "isActive": {
                          "type": "boolean"
                        },
                        "activeLoans": {
                          "type": "integer"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    },
                    "source": {
                      "type": "string",
                      "enum": [
                        "cache",
                        "database"
                      ]
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/pool/stats": {
      "get": {
        "summary": "Get aggregate lending pool statistics",
        "description": "Returns total deposits, utilization rate, current APY, and the number of active loans. Intended for the lender dashboard.\n",
        "tags": [
          "Pool"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Pool statistics retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PoolStatsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid Bearer token"
          }
        }
      }
    },
    "/pool/depositor/{address}": {
      "get": {
        "summary": "Get depositor portfolio for a wallet address",
        "description": "Returns deposit amount, pool share percentage, and estimated yield for the authenticated depositor. `address` must match the JWT wallet.\n",
        "tags": [
          "Pool"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "address",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Depositor's Stellar address (must match JWT)"
          }
        ],
        "responses": {
          "200": {
            "description": "Depositor portfolio retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepositorPortfolioResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid Bearer token"
          },
          "403": {
            "description": "address does not match authenticated wallet"
          }
        }
      }
    },
    "/pool/depositor/{address}/yield-history": {
      "get": {
        "summary": "Get per-depositor yield history",
        "description": "Returns a bounded time series of deposited value, current share value, and net yield for the authenticated depositor. Supports `days` query param (7, 30, or 90).\n",
        "tags": [
          "Pool"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "address",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "days",
            "schema": {
              "type": "integer",
              "enum": [
                7,
                30,
                90
              ],
              "default": 30
            }
          },
          {
            "in": "query",
            "name": "token",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Yield history retrieved successfully"
          },
          "401": {
            "description": "Missing or invalid Bearer token"
          },
          "403": {
            "description": "address does not match authenticated wallet"
          }
        }
      }
    },
    "/pool/{token}/share-price": {
      "get": {
        "summary": "Get current on-chain share price for a pool token",
        "description": "Returns the current share price from the LendingPool contract. The value is scaled by 1,000,000 (e.g. 1,050,000 = 1.05 ratio). Cached briefly to absorb burst requests.\n",
        "tags": [
          "Pool"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Token address"
          }
        ],
        "responses": {
          "200": {
            "description": "Share price retrieved successfully"
          },
          "401": {
            "description": "Missing or invalid Bearer token"
          }
        }
      }
    },
    "/pool/build-deposit": {
      "post": {
        "summary": "Build an unsigned deposit transaction",
        "description": "Builds an unsigned Soroban `deposit(provider, token, amount)` transaction XDR against the LendingPool contract. The frontend signs it with the user's wallet and submits via POST /api/pool/submit.\n",
        "tags": [
          "Pool"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "depositorPublicKey",
                  "token",
                  "amount"
                ],
                "properties": {
                  "depositorPublicKey": {
                    "type": "string",
                    "description": "Depositor's Stellar public key (must match JWT)"
                  },
                  "token": {
                    "type": "string",
                    "description": "Address of the token to deposit"
                  },
                  "amount": {
                    "type": "number",
                    "description": "Amount to deposit",
                    "example": 1000
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Unsigned transaction XDR returned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnsignedTransactionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Missing or invalid Bearer token"
          }
        }
      }
    },
    "/pool/build-withdraw": {
      "post": {
        "summary": "Build an unsigned withdraw transaction",
        "description": "Builds an unsigned Soroban `withdraw(provider, token, shares)` transaction XDR against the LendingPool contract. The frontend signs it with the user's wallet and submits via POST /api/pool/submit.\n",
        "tags": [
          "Pool"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "depositorPublicKey",
                  "token",
                  "amount"
                ],
                "properties": {
                  "depositorPublicKey": {
                    "type": "string",
                    "description": "Depositor's Stellar public key (must match JWT)"
                  },
                  "token": {
                    "type": "string",
                    "description": "Address of the token to withdraw"
                  },
                  "amount": {
                    "type": "number",
                    "description": "Amount (shares) to withdraw",
                    "example": 500
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Unsigned transaction XDR returned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnsignedTransactionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Missing or invalid Bearer token"
          }
        }
      }
    },
    "/pool/build-emergency-withdraw": {
      "post": {
        "summary": "Build an unsigned emergency withdraw transaction",
        "description": "Builds an unsigned Soroban `emergency_withdraw(provider, token, shares)` transaction XDR against the LendingPool contract. Bypasses the withdrawal cooldown. The frontend signs it with the user's wallet and submits via POST /api/pool/submit.\n",
        "tags": [
          "Pool"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "depositorPublicKey",
                  "token",
                  "shares"
                ],
                "properties": {
                  "depositorPublicKey": {
                    "type": "string",
                    "description": "Depositor's Stellar public key (must match JWT)"
                  },
                  "token": {
                    "type": "string",
                    "description": "Address of the token to withdraw"
                  },
                  "shares": {
                    "type": "number",
                    "description": "Amount (shares) to withdraw"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Unsigned transaction XDR returned"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Missing or invalid Bearer token"
          }
        }
      }
    },
    "/pool/submit": {
      "post": {
        "summary": "Submit a signed pool transaction",
        "description": "Submits a signed transaction XDR to the Stellar network for a pool deposit or withdrawal.\n",
        "tags": [
          "Pool"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "signedTxXdr"
                ],
                "properties": {
                  "signedTxXdr": {
                    "type": "string",
                    "description": "Signed transaction XDR"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transaction submitted and result returned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmittedTransactionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Missing or invalid Bearer token"
          }
        }
      }
    },
...
  },
...
}